Clean up to get the ChatFormatter out of the main gradio file handling the UI
