+++ tools/prompt_generator/prompt.py
@@ -1,27 +1,6 @@
 import gradio as gr
 
-from typing import (
-    AbstractSet,
-    cast,
-    Collection,
-    Dict,
-    Iterator,
-    List,
-    Literal,
-    Sequence,
-    TypedDict,
-    Union,
-)
-
-Role = Literal["system", "user", "assistant"]
-
-
-class Message(TypedDict):
-    role: Role
-    content: str
-
-
-Dialog = Sequence[Message]
+from prompt_utils import ChatFormat, Dialog
 
 def prompt_template_dropdown_listener(value):
     # This function will be called whenever the value of the dropdown changes
@@ -38,47 +17,26 @@ def prompt_template_dropdown_listener(value):
 
 def format_prompt_template_listener(system_prompt, user_prompt_1, assistant_response, user_prompt_2):
     if not user_prompt_1:
-        return {
-            ""
-        }
+        raise gr.Error("User prompt is mandatory.")
+
+    if not user_prompt_2 and assistant_response:
+        raise gr.Error("When the assistant message is set, the second user prompt is mandatory.")
+
+    if (user_prompt_2 and not assistant_response) or (not user_prompt_2 and not assistant_response):
+        raise gr.Error("When generating a multi turn prompt, the assistant message is mandatory.")
+
     dialog: Dialog = []
     if system_prompt: dialog.append({   "role": "system", "content": system_prompt, })
     dialog.append({   "role": "user", "content": user_prompt_1, })
     if assistant_response: dialog.append({   "role": "assistant", "content": assistant_response, })
     if user_prompt_2: dialog.append({   "role": "user", "content": user_prompt_2, })
 
-    return ChatFormat.encode_dialog_prompt(dialog)
-
-class ChatFormat:
-
-    @staticmethod
-    def encode_header( message: Message) -> List[str]:
-        prompts = []
-        prompts.append("<|start_header_id|>")
-        prompts.extend(message["role"])
-        prompts.append("<|end_header_id|>")
-        prompts.extend("\n\n")
-        return prompts
-
-    @staticmethod
-    def encode_message(message: Message) -> List[str]:
-        prompts = ChatFormat.encode_header(message)
-        prompts.extend(message["content"].strip())
-        prompts.append("<|eot_id|>")
-        return prompts
-
-    @staticmethod
-    def encode_dialog_prompt(dialog: Dialog) -> str:
-        prompts = ["<|begin_of_text|>"]
-        for message in dialog:
-            prompts.extend(ChatFormat.encode_message(message))
-        # Add the start of an assistant message for the model to complete.
-        prompts.extend(ChatFormat.encode_header({"role": "assistant", "content": ""}))
-        return {
-            prompt_output: "".join(prompts),
-            python_output: "",
-            hf_output: "",
-        }
+    return {
+        notice: gr.Markdown(elem_id="notice", visible=False, value="> Check the input prompts carefully before submitting."),
+        prompt_output: ChatFormat.encode_dialog_prompt(dialog),
+        python_output: "",
+        hf_output: "",
+    }
 
 
 with gr.Blocks() as demo:
@@ -90,17 +48,20 @@ with gr.Blocks() as demo:
             prompt_template = gr.Dropdown(["Single Turn", "Multi Turn"], label="Prompt Template", filterable=False)
 
             gr.Markdown("## Input Prompts")
-            system_prompt = gr.Textbox(label="System prompt", lines=2)
-            user_prompt_1 = gr.Textbox(label="User prompt", lines=2)
-            assistant_response = gr.Textbox(label="Assistant response", lines=2, visible=False, elem_id="assistant_response")
-            user_prompt_2 = gr.Textbox(label="User prompt", lines=2, visible=False, elem_id="user_prompt_2")
+            system_prompt = gr.Textbox(label="System prompt", lines=2, placeholder="Optional System prompt for the model")
+            user_prompt_1 = gr.Textbox(elem_id="user_prompt_1", label="User prompt *", lines=2, placeholder="Mandatory user prompt")
+            assistant_response = gr.Textbox(label="Assistant response", lines=2, visible=False, elem_id="assistant_response", placeholder="Mandatory assistant if prompt template is Multi Turn")
+            user_prompt_2 = gr.Textbox(label="User prompt *", lines=2, visible=False, elem_id="user_prompt_2", placeholder="Mandatory user prompt if prompt template is Multi Turn")
 
             prompt_template.input(prompt_template_dropdown_listener, prompt_template, [assistant_response, user_prompt_2])
 
+            submit = gr.Button("Submit")
 
         with gr.Column(scale=3, min_width=600):
             gr.Markdown("## Output")
             with gr.Tab("Preview"):
+
+                notice = gr.Markdown(elem_id="notice", value="> Check the input prompts carefully before submitting.", visible=False)
                 prompt_output = gr.Textbox(show_label=False, interactive=False, min_width=600, lines=30)
 
             with gr.Tab("Code"):
@@ -110,11 +71,6 @@ with gr.Blocks() as demo:
                     with gr.Tab("Hugging Face"):
                         hf_output = gr.Textbox(label="Using HF Transformers", interactive=False, min_width=600, lines=25)
 
-    system_prompt.change(format_prompt_template_listener, [system_prompt, user_prompt_1, assistant_response, user_prompt_2], [prompt_output, python_output, hf_output])
-    user_prompt_1.change(format_prompt_template_listener, [system_prompt, user_prompt_1, assistant_response, user_prompt_2], [prompt_output, python_output, hf_output])
-    assistant_response.change(format_prompt_template_listener, [system_prompt, user_prompt_1, assistant_response, user_prompt_2], [prompt_output, python_output, hf_output])
-    user_prompt_2.change(format_prompt_template_listener, [system_prompt, user_prompt_1, assistant_response, user_prompt_2], [prompt_output, python_output, hf_output])
-
-
+    submit.click(format_prompt_template_listener, [system_prompt, user_prompt_1, assistant_response, user_prompt_2], [notice, prompt_output, python_output, hf_output])
 
 demo.launch()
